requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

websockets==12.0

pyaudio==0.2.14

motor==3.3.2
pymongo==4.6.0

firebase-admin==6.3.0

python-dotenv==1.0.0

# Development/Testing (Optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1